% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.dataweighting.to.csl2.R
\name{apply.dataweighting.to.csl2}
\alias{apply.dataweighting.to.csl2}
\title{apply.dataweighting.to.csl2 reads in your observation.csl2 file and applys a reweighting factor and saves as a new .csl2 file.}
\usage{
apply.dataweighting.to.csl2(Path = "", weighting_factor,
  Observation_csl2_file = "Observations.csl2", Observation_label = "",
  Observation_out_filename = "Observation.csl2.0", fileEncoding = "")
}
\arguments{
\item{Path}{path to the directory containing the Observation.csl2 file.}

\item{weighting_factor}{- the multiplier for the effective sample sizes for the compositional data.}

\item{Observation_csl2_file}{the name of the configuration file containing the @observation blocks.}

\item{Observation_label}{the label of the observation you want to apply the weighting to.}

\item{Observation_out_filename}{the file name of the Observation.csl2 file to be saved after applying weighting.}

\item{fileEncoding}{Optional, allows the R-library to read in files that have been encoded in alternative UTF formats, see the manual for the error message that would indicate when to use this switch.}
}
\value{
a file named 'Observation_out_filename.out' in the path directory
}
\description{
This function reads in your observation.csl2 file and applys a reweighting factor and saves as a new .csl2 file, pretty much just trying to 
automate the dataweighting process. The problem with this method, is that it strips out all the comments from the original observation.csl2 file.
}
\author{
Craig Marsh
}
