% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn.in.tabular.R
\name{burn.in.tabular}
\alias{burn.in.tabular}
\title{burn.in.tabular used to return a list that has had the beginning cut off.}
\usage{
burn.in.tabular(tab_object, Row)
}
\arguments{
\item{tab_object}{casal2TAB object you want to burn-in}

\item{Row}{number to burn in from, note this is not the iteration but the row that corresponds to your iteration that you want to burn-in from. if keep > 1 then the iteration and row will be different}
}
\value{
a 'casal2TAB' object which has been manipulated.
}
\description{
returns a casal2TAB class that has been shortened to the ith iteration
}
\author{
Craig Marsh
}
