% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Casal2-package.R
\docType{package}
\name{Casal2-package}
\alias{Casal2-package}
\alias{Casal2}
\title{Casal2 R library}
\description{
A set of R functions for extracting and summarising output from
  Casal2 fisheries stock assessment model runs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/Casal2}
  \item Report bugs at \url{https://github.com/Casal2/CASAL2/issues}
}

}
\author{
Casal2 Development Team
}
