% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mpd_identifiability.R
\name{check_mpd_identifiability}
\alias{check_mpd_identifiability}
\title{check_mpd_identifiability}
\usage{
check_mpd_identifiability(cas2_mod, delta = .Machine$double.eps)
}
\arguments{
\item{model}{<casal2MPD> object that are generated from the extract.mpd() function. It expects the report of type
covariance_matrix to be present.}
}
\value{
a data frame or message about the model
}
\description{
Do an eigen decomposition of the hessian matrix to identify correlated estimated parameters
and parameters with high uncertainty. Small or zero eigenvalues (high condition number)
indicates ill-posedness, i.e. the parameter estimation problem does not have a unique solution.
This eigenvalue decomposition has been widely used in the estimation literature
}
\author{
C.Marsh
}
