% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_short_hand.R
\name{check_short_hand}
\alias{check_short_hand}
\title{Utility extract function}
\usage{
check_short_hand(x)
}
\arguments{
\item{x}{variable to check}
}
\description{
Utility extract function
}
\author{
Craig Marsh
}
