% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_short_hand.R
\name{checkShortHand}
\alias{checkShortHand}
\title{Utility extract function}
\usage{
checkShortHand(x)
}
\arguments{
\item{x}{variable to check}
}
\description{
Utility extract function
}
\author{
Casal2 Development Team
}
