% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.for.CPUE.R
\name{CV.for.CPUE}
\alias{CV.for.CPUE}
\title{CV.for.CPUE}
\usage{
CV.for.CPUE(year, cpue, f, plot.it = TRUE)
}
\arguments{
\item{year}{vector of years with CPUE indices}

\item{cpue}{CPUE indices}

\item{f}{degree of lowess smoothing (0 = no smoothing, 1 = maximum smoothing)}

\item{plot.it}{If TRUE, plot the index and the smoothed fit. Otherwise, return a dataframe of the year, index, smoothed fitted value, and cv)}
}
\value{
The function either plots the CPUE, together with a lowess line fitted to it, and returns the c.v. of the residuals to the fit. Or returns a dataframe of the lowess line fits and assocated c.v.s for each point.
}
\description{
This function is useful for deciding on a c.v. to be used with a CPUE series in a stock assessment model.
Originally written in Chris Francis's DataWeighting Package, this has been copied over and modified so that users can 
use this functionality with Casal2 models/output.
}
\note{
The idea is to fit a lowess line of "appropriate" smoothness to the CPUE data, and use the c.v. of the residuals from that fit in the stock assessment model. 
This is equivalent to saying that we expect the stock assessment model to fit these data as well as the smoother. You judge what is "appropriate" by visual examination of lines with different degrees of smoothing. 
This approach is recommended on p. 1132 of the following paper: Francis, R.I.C.C. (2011). Data weighting in statistical fisheries stock assessment models. Canadian Journal of Fisheries and Aquatic Sciences 68: 1124-1138.
}
\author{
Chris Francis
}
