% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleExponential.R
\name{DoubleExponential}
\alias{DoubleExponential}
\title{Calculate the Double-Exponential Selectivity used in Casal2}
\usage{
DoubleExponential(age_or_length, x0, x1, x2, y0, y1, y2, alpha = 1)
}
\arguments{
\item{age_or_length}{"vector<numeric>" A vector of ages and or lengths to calculate the selectivity over.}

\item{x0}{"numeric" See manual for formal definition of these parameters}

\item{x1}{"numeric"}

\item{x2}{"numeric"}

\item{y0}{"numeric"}

\item{y1}{"numeric"}

\item{y2}{"numeric"}

\item{alpha}{"numeric" The capping parameter of the selectivity, to move the max away from one}
}
\value{
"vector<numeric>" selectivity values over the age/length range and parameters supplied
}
\description{
This function calculates the Double-Exponential selectivity and should be used to plot up the form of the selectivity when estimating the parameters in the model.
}
\author{
Craig Marsh
}
