% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoubleNormal.R
\name{DoubleNormal}
\alias{DoubleNormal}
\title{Calculate the Double-Normal Selectivity used in Casal2}
\usage{
DoubleNormal(age_or_length, alpha = 1, mu, sigma_l, sigma_r)
}
\arguments{
\item{age_or_length}{"vector<numeric>" A vector of ages and or lengths to calculate the selectivity over.}

\item{alpha}{"numeric" The capping parameter of the selectivity, to move the max away from one}

\item{mu}{"numeric" Mean of the selectivity}

\item{sigma_l}{"numeric" left hand standard deviation of the selectivity}

\item{sigma_r}{"numeric" Right hand standard deviation of the selectivity}
}
\value{
"vector<numeric>" selectivity values over the age/length range and parameters supplied
}
\description{
This function calculates the Double-Normal selectivity and should be used to plot up the form of the selectivity when estimating the parameters in the model.
}
\author{
Craig Marsh
}

