% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{ess_bulk}
\alias{ess_bulk}
\title{ess_bulk}
\usage{
ess_bulk(sims)
}
\arguments{
\item{sims}{A 2D array \emph{without} warmup samples (iterations x chains),
or a vector when a single chain is supplied.}
}
\value{
A single numeric bulk-ESS value.
}
\description{
Compute the bulk effective sample size (bulk-ESS) for a single
  MCMC parameter using rank normalisation and split chains.
}
\references{
Vehtari et al. (2019) arXiv:1903.08008.
}
\author{
Aki Vehtari, Paul Bürkner (RStan); ported to casal2 R library.
}
