% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortHandFunctions.R
\name{expand_category_block}
\alias{expand_category_block}
\title{expand_category_block}
\usage{
expand_category_block(categories)
}
\arguments{
\item{categories}{string of categories}
}
\value{
a vector of strings
}
\description{
A utility function for expanding short hand syntax in @category blocks in casal2 config files
}
\examples{
\dontrun{
expand_category_block("stock")
expand_category_block("stock.male,female")
expand_category_block("stock.male,female.untagged,1990")
}
}
\author{
Craig Marsh
}
\keyword{internal}
