% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortHandFunctions.R
\name{expand_category_shorthand}
\alias{expand_category_shorthand}
\title{expand_category_shorthand}
\usage{
expand_category_shorthand(
  shorthand_categories,
  reference_categories,
  category_format = NULL
)
}
\arguments{
\item{shorthand_categories}{shorthand values to expand}

\item{reference_categories}{string of categories expanded from the @category block of a config. derived from expand_category_block}

\item{category_format}{the format defined in the @categories block}
}
\value{
a vector of strings
}
\description{
A utility function for expanding short hand syntax of categories used in subcommands throught casal2 config files
}
\examples{
\dontrun{
expand_category_shorthand(shorthand_categories ="format=*.EN.*", reference_categories = c("BOP.EN.male"))
expand_category_shorthand(shorthand_categories ="format=*.EN.*", reference_categories = c("BOP.EN.male", "BOP.EN.female"))
expand_category_shorthand(shorthand_categories ="format=*.EN.*", reference_categories = c("BOP.EN.male", "BOP.EN.female", "HAGU.EN.male", "HAGU.EN.female"))
expand_category_shorthand(shorthand_categories ="format = *.EN.+", reference_categories = c("BOP.EN.male", "BOP.EN.female"))
expand_category_shorthand(shorthand_categories ="format = *.EN.+", reference_categories = c("BOP.EN.male", "BOP.EN.female", "HAGU.EN.male", "HAGU.EN.female"))
expand_category_shorthand(shorthand_categories ="*", reference_categories = c("BOP.EN.male", "BOP.EN.female", "HAGU.EN.male", "HAGU.EN.female"))
expand_category_shorthand(shorthand_categories ="stock=BOP", reference_categories = c("BOP.EN.male", "BOP.EN.female", "HAGU.EN.male", "HAGU.EN.female"), category_format = "stock.area.sex")
}
}
\author{
Craig Marsh
}
\keyword{internal}
