% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortHandFunctions.R
\name{expand_shorthand_syntax}
\alias{expand_shorthand_syntax}
\title{expand_shorthand_syntax}
\usage{
expand_shorthand_syntax(syntax)
}
\arguments{
\item{syntax}{string of the syntax to expand}
}
\value{
a vector of strings
}
\description{
A utility function for expanding short hand syntax for based on - format=*.EN.*, label*5, and + syntax
}
\examples{
\dontrun{
# currently can handle these shorthand
expand_shorthand_syntax(syntax = "age_length")
expand_shorthand_syntax(syntax = "age_length * 8")
expand_shorthand_syntax(syntax ="age_length*8")
expand_shorthand_syntax(syntax ="age_length_BP*4 age_length_EN*4")
expand_shorthand_syntax(syntax = "1990:2000")
expand_shorthand_syntax(syntax ="age_length_BP*4,age_length_EN*4")
expand_shorthand_syntax(syntax ="0.75*3")
expand_shorthand_syntax(syntax ="0.75 * 3 ")
# cant deal with
expand_shorthand_syntax(syntax ="0.75*3 0.75*3")
}
}
\author{
Craig Marsh
}
\keyword{internal}
