% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.correlation.matrix.R
\name{extract.correlation.matrix}
\alias{extract.correlation.matrix}
\title{Utility extract.correlation.matrix function}
\usage{
extract.correlation.matrix(file, path = "", fileEncoding = "")
}
\arguments{
\item{file}{the name of the file containing the estimated covariance matrix ('mpd.out')}

\item{path}{(optional) the path to the file}

\item{fileEncoding}{(optional) allows the R library to read in files that have been encoded in alternative UTF formats. See the User Manual for the error message that would indicate when to use this switch.}
}
\value{
correlation matrix
}
\description{
This function reads in the 'mpd.out' file created in estimation mode
}
