% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.covariance.matrix.R
\name{extract.covariance.matrix}
\alias{extract.covariance.matrix}
\title{Utility extract.covariance.matrix function}
\usage{
extract.covariance.matrix(file, path = "", fileEncoding = "", quiet = FALSE)
}
\arguments{
\item{file}{the name of the file containing the estimated covariance matrix ('mpd.out')}

\item{path}{(optional) the path to the file}

\item{fileEncoding}{(optional) allows the R library to read in files that have been encoded in alternative UTF formats. See the User Manual for the error message that would indicate when to use this switch.}

\item{quiet}{suppress print or cat statements to screen.}
}
\value{
covariance matrix
}
\description{
This function reads in the 'mpd.out' file created in estimation mode
}
