% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.csl2.file.R
\name{extract.csl2.file}
\alias{extract.csl2.file}
\title{This function reads a Casal2 configuration file and returns a list object in R. Where each element is a command and subcommand from the configuration file}
\usage{
extract.csl2.file(file, path = "")
}
\arguments{
\item{file}{the name of the input file containing model configuration}

\item{path}{Optionally, the path to the file}
}
\description{
This function reads a Casal2 configuration file and returns a list object in R. Where each element is a command and subcommand from the configuration file
}
\author{
Craig Marsh
}

