% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.mcmc.R
\name{extract.mcmc}
\alias{extract.mcmc}
\title{extract.mcmc function for casal2 output}
\usage{
extract.mcmc(
  samples.file = "samples.1",
  objectives.file = "objectives.1",
  path = "",
  return_covariance = FALSE,
  fileEncoding = "",
  quiet = FALSE
)
}
\arguments{
\item{samples.file}{<string> the name of the input file containing the samples.file output by casal2}

\item{objectives.file}{<string> the name of the input file containing the objectives.file output by casal2}

\item{path}{Optional<string>, the path to the file}

\item{return_covariance}{Optional<bool>, Whether you want to extract the covariance matrix with the mcmc object?}

\item{fileEncoding}{Optional, allows the R-library to read in files that have been encoded in alternative UTF formats, see the manual for the error message that would indicate when to use this switch.}
}
\value{
a 'casal2MCMC' that can be integrated using the str() function.
}
\description{
An extract function that reads objective and sample output that are produced from a 'casal2 -m' model run. This funciton
also create a 'casal2.mcmc' class which can be used in plotting and summary functions.
}
\author{
C. Marsh
}
