% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Utility extract function}
\usage{
extract(file, path = "")
}
\arguments{
\item{file}{the name of the input file containing model output to extract}

\item{path}{Optionally, the path to the file}

\item{validate}{Optionally, validate the structure of each element in the report section}

\item{conversion}{Optionally, where possible, convert elements from list_element and data.frame into numeric}
}
\description{
Utility extract function
}
\author{
Dan Fu
}
