% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.starting.pars.R
\name{generate.starting.pars}
\alias{generate.starting.pars}
\title{generate.starting.pars Generates a parameter file that is formatted for -i input into Casal2.}
\usage{
generate.starting.pars(
  MPD = MPD,
  n_sim = 10,
  start_free_filename = "start_free.dat",
  path = "",
  all_uniform = FALSE,
  shrink = 0,
  quiet = T,
  write = TRUE
)
}
\arguments{
\item{MPD}{<casal2MPD> object that is generated from the extract() function.}

\item{n_sim}{the number of randomly generated sets of free parameters}

\item{start_free_filename}{= the filename of the file created}

\item{path}{Optionally, the path to location to write the output file}

\item{all_uniform}{= logical if TRUE draw from a uniform between bounds regardless of the prior distribution}

\item{shrink}{shrink bounds by a percent of the range of bounds to keep starting pars away from bounds (default = 0) and range (0-0.45)}

\item{quiet}{Optional, suppresses printing statements}

\item{write}{if false, do not write to the output file}
}
\value{
a file (with default name 'start_free.dat') in the 'path' directory and returns these values if the output is assigned to an object
}
\description{
This function takes an  Casal2 MPD  object with an estimate_summary report, and returns a free parameters file. Where each parameter is drawn from the prior defined in the estimate_summary block.
}
\author{
Alistair Dunn (derived from code by Craig Marsh)
}
