% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.casal2.unique_subcommands_list.R
\name{get.casal2.unique_subcommands_list}
\alias{get.casal2.unique_subcommands_list}
\title{Utility function}
\usage{
get.casal2.unique_subcommands_list()
}
\description{
Utility function
}
\author{
Craig Marsh
}
\keyword{internal}
