% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lines.R
\name{get.lines}
\alias{get.lines}
\title{Utility extract function}
\usage{
get.lines(
  lines,
  from = -1,
  to = -1,
  contains = "",
  starts.with = "",
  clip.to = "",
  clip.from = "",
  clip.to.match = "",
  clip.from.match = "",
  ...
)
}
\description{
Utility extract function
}
\author{
Dan Fu
}
\keyword{internal}
