% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abundance_observations.R
\name{has_abundance_observations}
\alias{has_abundance_observations}
\title{has_abundance_observations}
\usage{
has_abundance_observations(model, ...)
}
\arguments{
\item{model}{A \code{casal2MPD} object.}

\item{\dots}{Further arguments (currently unused).}
}
\value{
\code{TRUE} if at least one abundance/biomass observation report is present.
}
\description{
Check whether a Casal2 model output contains any abundance or
  biomass observations.
}
