% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catchabilities.R
\name{has_catchabilities}
\alias{has_catchabilities}
\title{has_catchabilities}
\usage{
has_catchabilities(model, ...)
}
\arguments{
\item{model}{A \code{casal2MPD} object.}

\item{\dots}{Further arguments (currently unused).}
}
\value{
\code{TRUE} if at least one \code{catchability} report is present.
}
\description{
Check whether a Casal2 model output contains any catchability reports.
}
