% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Increasing.R
\name{Increasing}
\alias{Increasing}
\title{Calculate the Increasing Selectivity used in Casal2}
\usage{
Increasing(age_or_length, alpha = 1, v, low, high)
}
\arguments{
\item{age_or_length}{"vector<numeric>" A vector of ages and or lengths to calculate the selectivity over.}

\item{alpha}{"numeric" The capping parameter of the selectivity, to move the max away from one}

\item{v}{"vector<numeric>" A vector of selectivities, must have a one for one relationship with age_or_length.}

\item{low}{"numeric" age or length at which selectivity is set = 0 below}

\item{high}{"numeric" age or length at which selectivity is set = alpha above}
}
\value{
"vector<numeric>" selectivity values over the age/length range and parameters supplied
}
\description{
This function calculates the Increasing selectivity and should be used to plot up the form of the selectivity when estimating the parameters in the model.
}
\author{
Craig Marsh
}
