% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/InverseLogistic.R
\name{InverseLogistic}
\alias{InverseLogistic}
\title{Calculate the Inverse-Logistic Selectivity used in Casal2}
\usage{
InverseLogistic(age_or_length, alpha = 1, a50, ato95)
}
\arguments{
\item{age_or_length}{"vector<numeric>" A vector of ages and or lengths to calculate the selectivity over.}

\item{alpha}{"numeric" The capping parameter of the selectivity, to move the max away from one}

\item{a50}{"numeric" age or length where selectivity 50\%}

\item{ato95}{"numeric" age or length difference where selectivity goes from 50\% - 95\% selective}
}
\value{
"vector<numeric>" selectivity values over the age/length range and parameters supplied
}
\description{
This function calculates the Inverse-Logistic selectivity and should be used to plot up the form of the selectivity when estimating the parameters in the model.
}
\author{
Craig Marsh
}

