% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simplex.R
\name{jacobiansimplex}
\alias{jacobiansimplex}
\title{jacobiansimplex
takes an unconstrained vector of length K - 1, and calculates the determinate of the jacobian, not this relates to density. You will need to log it for log-likelihood contributions}
\usage{
jacobiansimplex(yk)
}
\arguments{
\item{yk}{vector of length K - 1 takes unconstrained values}
}
\value{
xk unit vector of length K
}
\description{
jacobiansimplex
takes an unconstrained vector of length K - 1, and calculates the determinate of the jacobian, not this relates to density. You will need to log it for log-likelihood contributions
}
\details{
if you want the value that casal2 returns you need to log and take the negative, this function evaluates change in variable based on the denisty, but Casal2's objective function is with respect to negative log-likelihood.
}
