% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnifeEdge.R
\name{KnifeEdge}
\alias{KnifeEdge}
\title{Calculate the KnifeEdge Selectivity used in Casal2}
\usage{
KnifeEdge(age_or_length, alpha = 1, Edge)
}
\arguments{
\item{age_or_length}{"vector<numeric>" A vector of ages and or lengths to calculate the selectivity over.}

\item{alpha}{"numeric" The capping parameter of the selectivity, to move the max away from one}

\item{edge}{"numeric" age or length at which selectivity is 0 to the left or alpha to the right}
}
\value{
"vector<numeric>" selectivity values over the age/length range and parameters supplied
}
\description{
This function calculates the KnifeEdge selectivity and should be used to plot up the form of the selectivity when estimating the parameters in the model.
}
\author{
Craig Marsh
}
