% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LogisticProducing.R
\name{LogisticProducing}
\alias{LogisticProducing}
\title{Calculate the Logistic-Producing Selectivity used in Casal2}
\usage{
LogisticProducing(age_or_length, alpha = 1, a50, ato95, high, low)
}
\arguments{
\item{age_or_length}{"vector<numeric>" A vector of ages and or lengths to calculate the selectivity over.}

\item{alpha}{"numeric" The capping parameter of the selectivity, to move the max away from one}

\item{a50}{"numeric" age or length where selectivity 50\%}

\item{ato95}{"numeric" age or length difference where selectivity goes from 50\% - 95\% selective}

\item{high}{"numeric" age or length at which selectivity is set = alpha above}

\item{low}{"numeric" age or length at which selectivity is set = 0 below}
}
\value{
"vector<numeric>" selectivity values over the age/length range and parameters supplied
}
\description{
This function calculates the Logistic-Producing selectivity and should be used to plot up the form of the selectivity when estimating the parameters in the model.
}
\author{
Craig Marsh
}

