% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.data_frame_with_row_labs.R
\name{make.data_frame_with_row_labs}
\alias{make.data_frame_with_row_labs}
\title{Utility extract function}
\usage{
make.data_frame_with_row_labs(lines)
}
\description{
Utility extract function
}
\author{
C Marsh
}
\keyword{internal}
