% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.matrix_with_header.R
\name{make.matrix_with_header}
\alias{make.matrix_with_header}
\title{Utility extract function}
\usage{
make.matrix_with_header(lines)
}
\description{
create a matrix that has a header
}
\author{
Craig Marsh
}
\keyword{internal}
