% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.named_complete_vector.R
\name{make.named_complete_vector}
\alias{make.named_complete_vector}
\title{Utility extract function}
\usage{
make.named_complete_vector(lines)
}
\description{
Utility extract function
}
\author{
Dan Fu
}

