% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.data_frame_with_row_labs.R
\name{makeDataFrameRowLabels}
\alias{makeDataFrameRowLabels}
\title{Utility extract function}
\usage{
makeDataFrameRowLabels(lines)
}
\description{
Utility extract function
}
\author{
Casal2 Development Team
}
\keyword{internal}
