% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method.TA1.10.R
\name{Method.TA1.10}
\alias{Method.TA1.10}
\title{Method.TA1.10}
\usage{
Method.TA1.10(model, Observation_label, pe.cv.stg1, plotit = F, xlim = NULL,
  ylim = NULL, trial.pe.cvs = seq(0, 2, 0.05))
}
\arguments{
\item{model}{Casal2 output that is the result of a -r, -e run.}

\item{Observation_label}{Label of the observation you want to apply the weighting too.}

\item{pe.cv.stg1}{process-error c.v. assumed or estimated in stage 1 ("This can be extracted from the report")}

\item{xlim}{x-axis limits for the illustrative plot}

\item{ylim}{y-axis limits for the illustrative plot}

\item{trial.pe.cvs}{vector of trial stage-2 process-error c.v.s (the default value will usually be OK; if not, an error message is output)}

\item{plot.it}{If TRUE, plot the index and the smoothed fit. Otherwise, return a dataframe of the year, index, smoothed fitted value, and cv)}
}
\value{
Outputs a stage-2 process-error c.v. for the data set. Returns a 2-vector containing the process-error c.v. assumed or estimated in stage 1, and that estimated using TA1.10 for stage 2
}
\description{
This function is useful for deciding on the data weights of one or more at-age or at-length data sets with assumed lognormal error structure in a stock assessment.
Originally written in Chris Francis's DataWeighting Package, this has been copied over and modified so that users can use this 
functionality with Casal2 models/output.
}
\note{
Method TA1.10 is described in Appendix A of the following paper Francis, R.I.C.C. (2011). Data weighting in statistical fisheries stock assessment models. 
Canadian Journal of Fisheries and Aquatic Sciences 68: 1124-1138. (With corrections to the equation in Francis R.I.C.C. (2011) Corrigendum: Data weighting in statistical fisheries stock assessment models. 
Haven't exported this function because I don't think it works even in Chris's DataWeighting package.
}
\author{
Chris Francis
}
