% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method.TA1.8.R
\name{Method.TA1.8}
\alias{Method.TA1.8}
\title{Method.TA1.8}
\usage{
Method.TA1.8(model, observation_labels, plot.it = F, xlim = NULL,
  ylim = NULL)
}
\arguments{
\item{model}{Casal2 output that is the result of a -r, -e run.}

\item{plot.it}{If TRUE, plot the index and the smoothed fit. Otherwise, return a dataframe of the year, index, smoothed fitted value, and cv)}

\item{xlim}{x-axis limits for the illustrative plot}

\item{ylim}{y-axis limits for the illustrative plot}

\item{Observation_label}{Label of the observation you want to apply the weighting too.}
}
\value{
Outputs a stage-2 process-error c.v. for the data set. Returns a 2-vector containing the process-error c.v. assumed or estimated in stage 1, and that estimated using TA1.10 for stage 2
}
\description{
This function is useful for deciding on the data weights of one or more at-age or at-length data sets with assumed lognormal error structure in a stock assessment.
Originally written in Chris Francis's DataWeighting Package, this has been copied over and modified so that users can use this 
functionality with Casal2 models/output.
}
\note{
Method TA1.8 is described in Appendix A of the following paper Francis, R.I.C.C. (2011). Data weighting in statistical fisheries stock assessment models. 
Canadian Journal of Fisheries and Aquatic Sciences 68: 1124-1138. (With corrections to the equation in Francis R.I.C.C. (2011) Corrigendum: Data weighting in statistical fisheries stock assessment models.
}
\author{
Chris Francis
}
