% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd_derived_quantity.R
\name{mpd_derived_quantity}
\alias{mpd_derived_quantity}
\title{Utility function for summary}
\usage{
mpd_derived_quantity(report_list)
}
\description{
Utility function for summary
}
\author{
C Marsh
This is a utility function that will summarise a derived quantity report for a Casal2MPD class
}
