% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param.profile.by.cohort.R
\name{param.profile.by.cohort}
\alias{param.profile.by.cohort}
\title{Generate an objective score profile plot for a particular parameter and age compositional dataset as a consequence from a casal2 -p run.}
\usage{
param.profile.by.cohort(obs, param, extract_list, Rescale = F,
  ylab = "Objective contribution", xlab = "Parameter", y_min = -0.5,
  threshold = 10)
}
\arguments{
\item{obs}{"string" A report label that correspond to the age compositional observation that you would like to plot the profile for. Can be many observtions}

\item{param}{"string" The parameter label that the profile was run for. It should follow the syntax that is used in Casal2 that is, "block[label].parameter" e.g. process[Recruitment].b0}

\item{extract_list}{"string" The name of the R object that was extracted into R usign the extract() function.}

\item{Rescale}{"bool/logical" If true it will rescale the likelihoods to have a minimum on 0. This is done by subtracting off the minimum value of the likelihood series.}

\item{ylab}{"string" optional, y-axis label}

\item{xlab}{"string" optional, x-axis label}

\item{threshold}{"numeric" the minimum number of years a cohort is seen in the observation, for being used in the plot. if threshold = 5 then any cohort that has been sampled for a minimum of 5 years will be automatically plotted.}

\item{ymin}{"numeric" optional, minimum value parsed to the ylim parameter in plot()}
}
\description{
This function reads an extracted list from Casal2 and plots the likelihood/objective profile for an age compositional observation over the specified parameters range. This range is defined in the profile block of teh Casal2 configuration file. Casal2 calculates the profile steps as even intervals between the upper and lower bound by = (upper_bound - lower_bound) / (steps + 1). The plot will generate a line for each cohort and category. The reason I made this function was for the specific investigation of the natural mortality parameter M. Sometimes running a profile of aggregated compositional data on a parameter is not that informative. The aim of this plot is to show which cohorts are suggestive of certain parameters values, this will either create more confusion (most likely) or at least give the user something interesting to think about (perhaps cohort based processes).
}
\examples{
library(CASAL2)
Profile = extract("profile.log") ## this is an output file from a casal2 -p run
param.profile.by.cohort(obs = c("Tangaroa_propn_at_age_Jan"), param = "process[Instantaneous_Mortality].m(male)",extract_list = "Profile", Rescale = T,ylab = "Objective contribution", xlab = "M",threshold = 10)
}
\author{
Craig Marsh
}

