% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param.profile.R
\name{param.profile}
\alias{param.profile}
\title{Generate an objective score profile plot for a particular parameter as a consequence from a casal2 -p run}
\usage{
param.profile(obs, param, extract_list, Rescale = F,
  ylab = "Objective contribution", xlab = "Parameter", y_min = -0.5)
}
\arguments{
\item{obs}{"vector<string>" A vector of report labels that correspond to observations that you would like to plot the profile for. Can be many observtions}

\item{param}{"string" The parameter label that the profile was run for. It should follow the syntax that is used in Casal2 that is, "block[label].parameter" e.g. process[Recruitment].b0}

\item{extract_list}{"string" The name of the R object that was extracted into R usign the extract() function.}

\item{Rescale}{"bool/logical" If true it will rescale the likelihoods to have a minimum on 0. This is done by subtracting off the minimum value of the likelihood series.}

\item{ylab}{"string" optional, y-axis label}

\item{xlab}{"string" optional, x-axis label}

\item{ymin}{"numeric" optional, minimum value parsed to the ylim parameter in plot()}
}
\description{
This function reads an extracted list from Casal2 and plots the likelihood/objective profile for one or many observations over the specified parameters range. This range is defined in the profile block of the Casal2 configuration file. Casal2 calculates the profile steps as even intervals between the upper and lower bound by = (upper_bound - lower_bound) / (steps + 1)
}
\examples{
library(CASAL2)
Profile = extract("profile.log") ## this is an output file from a casal2 -p run
param.profile(obs = c("Tangaroa_propn_at_age_Jan","trawl_propn_at_age"), param = "process[Instantaneous_Mortality].m(male)",extract_list = "Profile", Rescale = T,ylab = "Objective contribution", xlab = "M")
}
\author{
Craig Marsh
}

