% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.derived_quantities.R
\name{plot.derived_quantities}
\alias{plot.derived_quantities}
\alias{plot.derived_quantities.casal2MPD}
\title{plot.derived_quantities default}
\usage{
plot.derived_quantities(model, type = "number", report_label = "", xlim,
  ylim, xlab, ylab, main, col, plot.it = T, ...)

\method{plot.derived_quantities}{casal2MPD}(model, type = "number",
  report_label = "", xlim, ylim, xlab, ylab, main, col, plot.it = T, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions.}

\item{type}{<string> whether numbers or scaled by B0.}

\item{report_label}{<string>}

\item{...}{remaining plotting functions.}
}
\value{
\code{NULL}
}
\description{
A plotting function to plot SSB's for the 'casal2TAB' and 'casal2MPD' objects.
}
\author{
C. Marsh
}
