% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fits.R
\name{plot.fits}
\alias{plot.fits}
\alias{plot.fits.casal2MPD}
\alias{plot.fits.casal2TAB}
\title{plot.fits default}
\usage{
plot.fits(model, report_label = "", plot.it = T, xlim, ylim, xlab, ylab,
  main, col, ...)

\method{plot.fits}{casal2MPD}(model, report_label = "", type = "fit",
  plot.it = T, xlim, ylim, xlab, ylab, main, col, ...)

\method{plot.fits}{casal2TAB}(model, report_label = "", type = "resid",
  plot.it = T, xlim, ylim, xlab, ylab, main, col, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions.}

\item{report_label}{<string>}

\item{plot.it}{Whether to generate a default plot or return the values as a matrix.}

\item{...}{remaining plotting options}

\item{type}{Whether to plot an observed vs expected (fit) or plot the residuals (resid)}
}
\value{
A plot of derived quantities over time if plot.it = T, if plot.it = F it will return a matrix of derived quantities.

\code{NULL}

\code{NULL}
}
\description{
A plotting function to fits to Casal2 observations from a model run.
}
\examples{
library(casal2)
# plotting Standard Output
data <- extract.mpd(file = system.file("extdata", "estimate.log", package="casal2"))
names(data)
par(mfrow = c(1,2))
plot.fits(model = data, report_label = "Tangaroa_propn_at_age_Aug")
plot.fits(model = data, report_label = "wcsiTRLcpue")
# if you are unhappy with the default plotting you can use plot.it = FALSE and create a plot of your own.
Tangaroa_fits = plot.fits(model = data, report_label = "Tangaroa_propn_at_age_Aug", plot.it = FALSE)
# plotting Tabular Output
tab <- extract.tabular(file = system.file("extdata", "tabular_report.out", package="casal2"))
names(tab)
plot.fits(model = tab, report_label = "Tangaroa_propn_at_age_Aug")
par(mfrow = c(1,1))
plot.fits(model = tab, report_label = "wcsiTRLcpue")
}
\author{
Craig Marsh
}
