% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.prior.R
\name{plot.prior}
\alias{plot.prior}
\title{plot.prior plot that could be assumed in a Casal2 model}
\usage{
\method{plot}{prior}(type = "lognormal", mu, sd, cv, A = 0, B = 1,
  bounds, xlim, label = F, xlab = "x", ylab = "Density", logx = F,
  add = F, dump = F, n = 201, ...)
}
\value{
plot of a priors
}
\description{
A plotting a prior, to visulise
}
\author{
A.Dunn
}
