% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.recruitment.R
\name{plot.recruitment}
\alias{plot.recruitment}
\alias{plot.recruitment.casal2MPD}
\alias{plot.recruitment.casal2TAB}
\title{plot.recruitment plot the stock recruitment relationship from a Casal2 model output.}
\usage{
plot.recruitment(model, report_label = "", xlim, ylim, xlab, ylab, main,
  col, ...)

\method{plot.recruitment}{casal2MPD}(model, report_label = "",
  xlim = NULL, ylim = NULL, xlab = NULL, ylab = NULL,
  main = NULL, col = NULL, ...)

\method{plot.recruitment}{casal2TAB}(model, report_label = "", xlim,
  ylim, xlab, ylab, main, col, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}

\item{report_label}{<string> the report label containing recruitment process}

\item{...}{remaining plotting options}
}
\value{
generate a plot of SSBy vs Ry with the assumed Stock recruitment relationship through it.

\code{NULL}

\code{NULL}
}
\description{
A plotting function to plot the stock recruitment relationship for the 'casal2TAB' and 'casal2MPD' objects.
}
\author{
Craig Marsh
}
