% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.selectivities.R
\name{plot.selectivities}
\alias{plot.selectivities}
\alias{plot.selectivities.casal2MPD}
\title{plot.selectivities plot selectivities from a Casal2 model.}
\usage{
plot.selectivities(
  model,
  report_labels = c(""),
  xlim,
  ylim,
  xlab,
  ylab,
  main,
  col,
  ...
)

\method{plot.selectivities}{casal2MPD}(
  model,
  report_labels = c(""),
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col = NULL,
  ...
)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}

\item{...}{remaining plotting options}

\item{report_label}{vector<string>}
}
\value{
generate a plot by age or length

\code{NULL}
}
\description{
A plotting function to plot selectivities for the 'casal2TAB' and 'casal2MPD' objects.
}
\author{
Craig Marsh
}
