% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.prior.R
\name{plot_prior}
\alias{plot_prior}
\title{plot_prior}
\usage{
plot_prior(
  type = "lognormal",
  mu,
  sd,
  cv,
  A = 0,
  B = 1,
  beta = 0,
  bounds,
  xlim,
  label = F,
  xlab = "x",
  ylab = "Density",
  logx = F,
  add = F,
  dump = F,
  ...
)
}
\arguments{
\item{type}{The prior type}

\item{mu}{mean of the distribution for the prior}

\item{sd}{standard deviation (where required) of the distribution for the prior}

\item{cv}{CV (where required) of the distribution for the prior}

\item{A}{The A parameter for the beta prior}

\item{B}{The B parameter for the beta prior}

\item{beta}{The beta parameter for the prior}

\item{bounds}{The bounds of the prior}

\item{xlim}{The x-axis limits for the plot}

\item{label}{TRUE/FALSE (default = FALSE) display the name of the prior}

\item{logx}{TRUE/FALSE (default = FALSE) to plot the xaxis in log-space}

\item{add}{TRUE/FALSE (default = FALSE) to add to an existing plot}

\item{dump}{TRUE/FALSE (default = FALSE) to dump the coordinates of the plotted line}
}
\description{
Plot priors available in Casal2
}
\author{
A. Dunn
}
