% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadSimulatedData.R
\name{ReadSimulatedData}
\alias{ReadSimulatedData}
\title{Read in multiple sets of Simualted data for a single observation}
\usage{
ReadSimulatedData(filename, path = "")
}
\arguments{
\item{filename}{the name of simulated obs for an observation. For example if you generated 100 sets of simulated observations named "SubAntarticObs". Casal2 will generate 100 of these with the following extensions SubAntarticObs.001, SubAntarticObs.002, SubAntarticObs.003.,,, SubAntarticObs.100. filename = SubAntarticObs.}

\item{path}{Optionally, the path to the file, default is current working directory.}
}
\description{
This function reads in a set of simulated observations generated from Casal2 in simulation mode. These functions read in all the simulated obs as a list, for visualising and summarising in R
}
\author{
Craig Marsh
}

