% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.compositional.data.R
\name{reformat.compositional.data}
\alias{reformat.compositional.data}
\title{Reformat Casal2 composition observation data so it is in the same format as the legacy CASAL observation.}
\usage{
reformat.compositional.data(model, report_label)
}
\arguments{
\item{model}{The R object that has been extracted using the extract() function.}

\item{report_label}{<string> The label of the report for the observation to convert}
}
\description{
This function will take a composition observation data that has been used in or generated by Casal2 and re-format it
so that it has the same structure as a CASAL reported composition observation data. The purpose of this function is to
reformat the Casal2 observations so they can be used as input for functions on CASAL observation data, such as
Chris Francis' Data Weighting function.
}
\author{
Craig Marsh
}
