% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove.first.words.R
\name{remove.first.words}
\alias{remove.first.words}
\title{Utility extract function}
\usage{
remove.first.words(string, words = 1)
}
\description{
Utility extract function
}
\author{
Dan Fu
}

