% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.last.words.R
\name{remove.last.words}
\alias{remove.last.words}
\title{Utility extract function}
\usage{
remove.last.words(string, words = 1)
}
\description{
Utility extract function
}
\author{
Dan Fu
}
\keyword{internal}
