% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simplex.R
\name{restoresimplex}
\alias{restoresimplex}
\title{restoresimplex
takes an unconstrained vector of length K - 1, and calculates a unit vector of length K}
\usage{
restoresimplex(yk)
}
\arguments{
\item{yk}{vector of length K - 1 takes unconstrained values}
}
\value{
xk unit vector of length K
}
\description{
restoresimplex
takes an unconstrained vector of length K - 1, and calculates a unit vector of length K
}
