% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlnorm.R
\name{Rlnorm}
\alias{Rlnorm}
\title{Utility extract function}
\usage{
Rlnorm(n, mu, cv)
}
\arguments{
\item{n}{number of samples to draw}

\item{mu}{mean in normal space}

\item{cv}{cv in normal space}
}
\value{
n randomly generated values from a lognormal distribution
}
\description{
Utility extract function
}
\author{
Craig marsh
}
