% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.class.R
\name{set.class}
\alias{set.class}
\title{Set the class of an R object}
\usage{
set.class(object, new.class)
}
\arguments{
\item{object}{R object to modify}

\item{new.class}{character string naming the new class to prepend}
}
\description{
Set the class of an R object
}
\keyword{internal}
