% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simplex.R
\name{simplex}
\alias{simplex}
\title{simplex
takes a unit vector of length K and converts to unconstrained K - 1 vector}
\usage{
simplex(xk, sum_to_one = TRUE)
}
\arguments{
\item{xk}{vector of length K - 1 takes unconstrained values}

\item{sum_to_one}{whether to rescale xk so it sums to one}
}
\value{
vector of length K - 1 unconstrained values
}
\description{
simplex
takes a unit vector of length K and converts to unconstrained K - 1 vector
}
