% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.to.vector.of.numbers.R
\name{string.to.vector.of.numbers}
\alias{string.to.vector.of.numbers}
\title{Utility extract function}
\usage{
string.to.vector.of.numbers(string)
}
\description{
Utility extract function
}
\author{
Dan Fu
}
\keyword{internal}
