% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_estimate_summary.R
\name{summarise_estimate_summary}
\alias{summarise_estimate_summary}
\title{Utility summarise_estimate_values function}
\usage{
summarise_estimate_summary(report_object)
}
\description{
used in the summarise function for casal2MPD
}
\keyword{internal}
