% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_estimate_values.R
\name{summarise_estimate_values}
\alias{summarise_estimate_values}
\title{Utility summarise_estimate_values function}
\usage{
summarise_estimate_values(report_object)
}
\description{
used in the summarise function for casal2MPD
}
\keyword{internal}
