% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_process.R
\name{summarise_process}
\alias{summarise_process}
\title{Utility summarise_process function}
\usage{
summarise_process(report_object)
}
\description{
used in the summarise function for casal2MPD
}
\author{
Craig Marsh
}
\keyword{internal}
