% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_process_retained.R
\name{summarise_process_retained}
\alias{summarise_process_retained}
\title{Utility summarise_process_retained function}
\usage{
summarise_process_retained(report_object)
}
\description{
used in the summarise function for casal2MPD
}
\author{
Craig Marsh
}
\keyword{internal}
