% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_warnings_encounted.R
\name{summarise_warnings_encounted}
\alias{summarise_warnings_encounted}
\title{Utility summarise_warnings_encounted function}
\usage{
summarise_warnings_encounted(report_object)
}
\description{
used in the summarise function for casal2MPD
}
\author{
Craig Marsh
}
\keyword{internal}
