% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.default}
\alias{summary.default}
\alias{summary.casal2MPD}
\alias{summary.casal2MCMC}
\alias{summary.casal2TAB}
\title{Utility summary function}
\usage{
summary.default(model)

\method{summary}{casal2MPD}(model)

\method{summary}{casal2MCMC}(model)

\method{summary}{casal2TAB}(model)
}
\arguments{
\item{model}{The casal2TAB object}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
This function summarises information in 'casal2MCMC', 'casal2TAB', and 'casal2MPD' objects.
}
\author{
C. Marsh
}
