% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.comparisons.R
\name{transform.comparisons}
\alias{transform.comparisons}
\title{Utility function for observation related functionality}
\usage{
\method{transform}{comparisons}(Data)
}
\description{
Utility function for observation related functionality
}
\author{
Craig marsh
}
\keyword{internal}
