% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.csl2.file.R
\name{write.csl2.file}
\alias{write.csl2.file}
\title{Model configuration write function}
\usage{
write.csl2.file(object, file, path = "")
}
\arguments{
\item{object}{An R list object that follows the same structure that extract.csl2.file would produce}

\item{file}{Optionally, the file name}

\item{path}{Optionally, the path to ouput the file}
}
\description{
This function will write a Casal2 configuration file based on a list object in R. Ususally this function will be used once a model has been read into R using extract.csl2.file and modified. This function will then print our the configuration to a new file where it can be re run into Casal2
}
\author{
Craig Marsh
}

