% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.csl2.file.R
\name{extract.csl2.file}
\alias{extract.csl2.file}
\title{Utility extract function}
\usage{
extract.csl2.file(file, path = "")
}
\arguments{
\item{file}{the name of the input file containing model configuration}

\item{path}{Optionally, the path to the file}
}
\description{
Utility extract function
}
\author{
Craig Marsh
}

