% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.starting.pars.R
\name{generate.starting.pars}
\alias{generate.starting.pars}
\title{generate.starting.pars Generates a parameter file that is formatted for -i input into Casal2.}
\usage{
generate.starting.pars(
  path = "",
  estimation_csl2_file = "Estimation.csl2",
  n_sim = 10,
  par_file_name = "starting_pars.out",
  all_uniform = FALSE,
  fileEncoding = "",
  quiet = T
)
}
\arguments{
\item{path}{Optionally, the path to the file}

\item{estimation_csl2_file}{the name of the configuration file containing the @estimate blocks.}

\item{n_sim}{the number of random samples you want.}

\item{par_file_name}{= the filename of the file created}

\item{all_uniform}{= logical if TRUE draw from a uniform between bounds regardless of prior distribution}

\item{fileEncoding}{Optional, allows the R-library to read in files that have been encoded in alternative UTF formats, see the manual for the error message that would indicate when to use this switch.}

\item{quiet}{Optional, supresses printing statements}
}
\value{
a file named 'parms.out' in the path directory
}
\description{
This function reads a Casal2 estimation configuration file and returns a par file. Where each parameter is drawn from the prior defined in an @estimate block.
}
\author{
Craig Marsh
}
