% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.derived_quantities.R
\name{plot.derived_quantities}
\alias{plot.derived_quantities}
\alias{plot.derived_quantities.casal2MPD}
\alias{plot.derived_quantities.casal2TAB}
\title{plot.derived_quantities default}
\usage{
plot.derived_quantities(
  model,
  report_label = "",
  type = "number",
  xlim,
  ylim,
  xlab,
  ylab,
  main,
  col,
  plot.it = T,
  ...
)

\method{plot.derived_quantities}{casal2MPD}(
  model,
  report_label = "",
  type = "number",
  xlim,
  ylim,
  xlab,
  ylab,
  main,
  col,
  plot.it = T,
  ...
)

\method{plot.derived_quantities}{casal2TAB}(
  model,
  report_label = "",
  type = "number",
  xlim,
  ylim,
  xlab,
  ylab,
  main,
  col,
  plot.it = T,
  ...
)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions.}

\item{report_label}{<string>}

\item{type}{<string> whether numbers or scaled by B0.}

\item{plot.it}{Whether to generate a default plot or return the values as a matrix.}

\item{...}{remaining plotting options}
}
\value{
A plot of derived quantities over time if plot.it = T, if plot.it = F it will return a matrix of derived quantities.

\code{NULL}

\code{NULL}
}
\description{
A plotting function to plot Derived Quantities for the 'casal2TAB' and 'casal2MPD' objects.
}
\author{
Craig Marsh
}
