% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fits.R
\name{plot.fits}
\alias{plot.fits}
\alias{plot.fits.casal2MPD}
\alias{plot.fits.casal2TAB}
\title{plot.fits default}
\usage{
plot.fits(
  model,
  report_label = "",
  plot.it = T,
  xlim,
  ylim,
  xlab,
  ylab,
  main,
  col,
  ...
)

\method{plot.fits}{casal2MPD}(
  model,
  report_label = "",
  type = "fit",
  plot.it = T,
  xlim,
  ylim,
  xlab,
  ylab,
  main,
  col,
  ...
)

\method{plot.fits}{casal2TAB}(
  model,
  report_label = "",
  type = "resid",
  plot.it = T,
  xlim,
  ylim,
  xlab,
  ylab,
  main,
  col,
  ...
)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract() functions.}

\item{report_label}{<string>}

\item{plot.it}{Whether to generate a default plot or return the values as a matrix.}

\item{xlim}{The x-axis limits for the plot}

\item{ylim}{The y-axis limits for the plot}

\item{xlab}{The x-axis label for the plot}

\item{ylab}{The y-axis label for the plot}

\item{main}{The title of the plot}

\item{col}{The color used for plot}

\item{...}{remaining plotting options}
}
\value{
A plot of derived quantities over time if plot.it = TRUE, if plot.it = FALSE it will return a matrix of derived quantities.

\code{NULL}

\code{NULL}
}
\description{
A plotting function to fits to Casal2 observations from a model run.
}
\examples{
\donttest{
library(casal2)
# plotting Standard Output
data <- extract.mpd(file = system.file("extdata", "estimate.out", package="casal2"))
names(data)
par(mfrow = c(1,2))
plot.fits(model = data, report_label = "westF_at_age")
plot.fits(model = data, report_label = "eastF_at_age")
# to create a user-specified plot, use plot.it = FALSE.
Tangaroa_fits <- plot.fits(model = data, report_label = "eastF_at_age", plot.it = FALSE)
}
}
\author{
Craig Marsh
}
