% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pressure.R
\name{plot.pressure}
\alias{plot.pressure}
\alias{plot.pressure.casal2MPD}
\title{plot.pressure plot fishing pressure if there has been an exploitation process reported.}
\usage{
plot.pressure(model, report_label = "", xlim, ylim, xlab, ylab, main, col,
  plot.it = T, ...)

\method{plot.pressure}{casal2MPD}(model, report_label = "", xlim = NULL,
  ylim = NULL, xlab = NULL, ylab = NULL, main = NULL, col = NULL,
  plot.it = T, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}

\item{report_label}{<string>}

\item{...}{remaining plotting functions.}
}
\value{
\code{NULL}
}
\description{
A plotting function to plot fishing presuure (U's )for the 'casal2TAB' and 'casal2MPD' objects.
}
\examples{
library(casal2)
data <- extract.mpd(file = system.file("extdata", "MPD.log", package="casal2"))
names(data)
plot.pressure(model = data, report_label = "exploit", col = c("black", "red"))
# if you are unhappy with the default plotting you can use plot.it = FALSE and create a plot of your own.
Fish_pressure = plot.pressure(model = data, report_label = "exploit", plot.it = FALSE)
}
\author{
C. Marsh
}
