% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pressure.R
\name{plot.pressure}
\alias{plot.pressure}
\alias{plot.pressure.casal2MPD}
\alias{plot.pressure.casal2TAB}
\title{plot.pressure plot fishing pressure if there has been an exploitation process reported.}
\usage{
plot.pressure(model, report_label = "", xlim, ylim, xlab, ylab, main,
  col, plot.it = T, ...)

\method{plot.pressure}{casal2MPD}(model, report_label = "",
  xlim = NULL, ylim = NULL, xlab = NULL, ylab = NULL,
  main = NULL, col = NULL, plot.it = T, ...)

\method{plot.pressure}{casal2TAB}(model, report_label = "", xlim, ylim,
  xlab, ylab, main, col, plot.it = T, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}

\item{report_label}{<string>}

\item{plot.it}{Whether to generate a default plot or return the values as a matrix.}

\item{...}{remaining plotting options}
}
\value{
generate a plot over time if plot.it = T, if plot.it = F it will return a matrix of values.

\code{NULL}

\code{NULL}
}
\description{
A plotting function to plot fishing presuure (U's )for the 'casal2TAB' and 'casal2MPD' objects.
}
\author{
Craig Marsh
}
