% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ycs.R
\name{plot.ycs}
\alias{plot.ycs}
\alias{plot.ycs.casal2MPD}
\title{plot.ycs plot Year Class Strengths from a Casal2 model.}
\usage{
plot.ycs(model, report_label = "", xlim, ylim, xlab, ylab, main, col,
  plot.it = T, ...)

\method{plot.ycs}{casal2MPD}(model, report_label = "", xlim = NULL,
  ylim = NULL, xlab = NULL, ylab = NULL, main = NULL, col = NULL,
  plot.it = T, ...)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB> object that are generated from one of the extract.mpd() and extract.tabular() functions.}

\item{report_label}{<string>}

\item{...}{remaining plotting functions.}
}
\value{
\code{NULL}
}
\description{
A plotting function to plot YCS for the 'casal2TAB' and 'casal2MPD' objects.
}
\author{
C. Marsh
}
