% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_estimate_values.R,
%   R/summarise_process.R
\name{summarise_process}
\alias{summarise_process}
\alias{summarise_process}
\title{Utility summarise_estimate_values function}
\usage{
summarise_process(report_object)

summarise_process(report_object)
}
\description{
used in the summarise function for casal2MPD

used in the summarise function for casal2MPD
}
\author{
Craig Marsh

Craig Marsh
}
\keyword{internal}
