% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.default}
\alias{summary.default}
\alias{summary.casal2MPD}
\title{summary default}
\usage{
summary.default(model)

\method{summary}{casal2MPD}(model)
}
\arguments{
\item{model}{<casal2MPD, casal2TAB, casal2MCMC> object that are generated from one of the extract() functions.}
}
\value{
\code{NULL}
}
\description{
A summary function for 'casal2MCMC' 'casal2TAB' and 'casal2MPD' objects.
}
\author{
C. Marsh
}
