% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.comparisons.R
\name{transform.comparisons}
\alias{transform.comparisons}
\title{Utility extract function}
\usage{
\method{transform}{comparisons}(Data)
}
\description{
Utility extract function
}
\author{
Craig marsh
}
