% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.csl2.file.R
\name{write.csl2.file}
\alias{write.csl2.file}
\title{Model configuration write function}
\usage{
write.csl2.file(object, file, path = "")
}
\arguments{
\item{object}{An R list object that follows the same structure that extract.csl2.file would produce}

\item{file}{The file name}

\item{path}{The path to output the file (optional)}
}
\description{
This function will write a Casal2 configuration file based on a list object in R. Ususally this function
is used once a model has been read into R using extract.csl2.file and modified. This function will then
output the configuration to a new file where it can be rerun in Casal2
}
\author{
Craig Marsh
}
